Callback signature for write().

+++ xml.h
@@ -1,5 +1,5 @@
 #include <stddef.h>
 #ifndef XML_h
 #define XML_h
-int writer(void *, size_t, size_t, void*);
+size_t writer(void *, size_t, size_t, void*);
 #endif
